Quora question Pairs
The goal of this project is to build a machine learning model that can predict whether two given questions on Quora are duplicates or not. By using NLP techniques, we aim to analyze and compare the semantic meaning of pairs of questions and classify them as either similar or dissimilar.